+++ contracts/functional-boundary-baseline.json
@@ -8,7 +8,6 @@
     { "path": "tests/functional/guards_batch/factory_request_batch_long_test.go", "sha256": "0502d6efc18a2b6f0fc687ef4ff371c5b2109235485bcebc2bb22378f8728e8e" },
     { "path": "tests/functional/internal/support/api_server.go", "sha256": "7248cf8ff749f0db662a7d18d8570652ed02f9dbb846d97522bdcd380115df2c" },
     { "path": "tests/functional/internal/support/cmd/browser_api_harness/main.go", "sha256": "2037a06c5d7d7f54b5203d6f23f9702fcd9b988f622e2e1f2715db77362c45e2" },
-    { "path": "tests/functional/replay_contracts/javascript_recording_contract_test.go", "sha256": "dd55dfe16a2fff6dc6c01664012dec3a6edf801b996d432973c58eefdf138588" },
     { "path": "tests/functional/replay_contracts/replay_event_stream_artifact_smoke_long_test.go", "sha256": "12a38d450728ae3a15cb1cb8c51c7ce2e33fa0448bb8ca0101ac10768cabe31f" },
     { "path": "tests/functional/replay_contracts/replay_factory_only_serialization_smoke_long_test.go", "sha256": "07baabc839e94f4a5a32c1a00a8e1c6ff3fa87875771d24ec5dce843edc6f4d6" },
     { "path": "tests/functional/replay_contracts/replay_live_helpers_test.go", "sha256": "050ec1a39d91936626630d626f3b07d1162e718651195d1252b33a3d079039fe" },
